Moved clever_agg in reinsurance.py
